Move to :core:testing.
