--- Compose-Flight-Search-App/app/src/androidTest/kotlin/com/example/flightsearchapp/testing/model/ui/FlightModelTestData.kt
@@ -1,21 +1,0 @@
-package com.example.flightsearchapp.testing.model.ui
-
-import com.example.flightsearchapp.ui.model.Flight
-
-val flightTestData = Flight(
-    id = "departure_arrive",
-    departureIataCode = "ICN",
-    departureName = "Incheon",
-    arriveIataCode = "CDG",
-    arriveName = "Paris airport",
-    isBookmarked = false,
-)
-
-fun Flight.bookMarkToggle() = Flight(
-    id,
-    departureName,
-    departureIataCode,
-    arriveIataCode,
-    arriveName,
-    !isBookmarked,
-)
