+++ Compose-Flight-Search-App/core/testing/src/main/kotlin/com/example/flightsearchapp/core/testing/fake/ui/FlightModelTestData.kt
@@ -1,0 +1,22 @@
+package com.example.flightsearchapp.core.testing.fake.ui
+
+import com.example.flightsearchapp.core.model.Flight
+
+val flightTestData = Flight(
+    id = "departure_arrive",
+    departureIataCode = "ICN",
+    departureName = "Incheon",
+    arriveIataCode = "CDG",
+    arriveName = "Paris airport",
+    isBookmarked = false,
+)
+
+fun Flight.bookMarkToggle() =
+    Flight(
+        id,
+        departureName,
+        departureIataCode,
+        arriveIataCode,
+        arriveName,
+        !isBookmarked,
+    )
